x64 PE file mapping added (experimental) PE file scraper bug fixed (Only Workaround) More analysis verbose output.

+++ src/common.go
@@ -55,7 +55,7 @@ func verbose(str string, status string) {
 	}
 }
 
-func _verbose(str string, value int32) {
+func _verbose(str string, value uint64) {
 
 	if target.verbose {
 		BoldYellow.Print("[*] ")
